pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ssh-audio-play"
version = "0.1.0"
dependencies = ["python-decouple"]
description = "A tool to play audio files over SSH."
